+++ src/middlewared/middlewared/api/v25_10_0/__init__.py
@@ -1,6 +1,5 @@
 from .acl import *
 from .acme_dns_authenticator import *
-from .acme_registration import *
 from .alert import *
 from .alertservice import *
 from .api_key import *
